+++ zcash_primitives/src/transaction/components/amount.rs
@@ -23,6 +23,18 @@ pub const DEFAULT_FEE: Amount = Amount(1000);
 #[derive(Clone, Copy, Debug, PartialEq, PartialOrd, Eq, Ord)]
 pub struct Amount(i64);
 
+impl memuse::DynamicUsage for Amount {
+    #[inline(always)]
+    fn dynamic_usage(&self) -> usize {
+        0
+    }
+
+    #[inline(always)]
+    fn dynamic_usage_bounds(&self) -> (usize, Option<usize>) {
+        (0, Some(0))
+    }
+}
+
 impl Amount {
     /// Returns a zero-valued Amount.
     pub const fn zero() -> Self {